# Conflicts:
#	core/src/integration-test/java/com/scalar/db/storage/dynamo/ConsensusCommitAdminIntegrationTestWithDynamo.java
#	core/src/integration-test/java/com/scalar/db/storage/dynamo/DynamoAdminIntegrationTest.java
#	core/src/integration-test/java/com/scalar/db/storage/jdbc/ConsensusCommitAdminIntegrationTestWithJdbcDatabase.java
#	core/src/integration-test/java/com/scalar/db/storage/jdbc/JdbcAdminIntegrationTest.java
#	core/src/integration-test/java/com/scalar/db/storage/jdbc/JdbcAdminTestUtils.java
#	core/src/integration-test/java/com/scalar/db/storage/multistorage/MultiStorageAdminTestUtils.java
#	core/src/integration-test/java/com/scalar/db/transaction/jdbc/JdbcTransactionAdminIntegrationTest.java
#	core/src/main/java/com/scalar/db/api/Admin.java
#	core/src/main/java/com/scalar/db/storage/cosmos/CosmosAdmin.java
#	core/src/main/java/com/scalar/db/storage/dynamo/DynamoAdmin.java
#	core/src/main/java/com/scalar/db/storage/jdbc/JdbcAdmin.java
#	core/src/test/java/com/scalar/db/storage/cosmos/CosmosAdminTestBase.java
#	core/src/test/java/com/scalar/db/storage/dynamo/DynamoAdminTestBase.java
#	core/src/test/java/com/scalar/db/storage/dynamo/DynamoConfigTest.java
#	core/src/test/java/com/scalar/db/storage/jdbc/JdbcAdminTestBase.java
#	core/src/test/java/com/scalar/db/storage/jdbc/JdbcConfigTest.java
#	integration-test/src/main/java/com/scalar/db/api/DistributedStorageAdminIntegrationTestBase.java
#	integration-test/src/main/java/com/scalar/db/api/DistributedTransactionAdminIntegrationTestBase.java
#	schema-loader/src/test/java/com/scalar/db/schemaloader/SchemaOperatorTest.java
#	server/src/integration-test/java/com/scalar/db/server/ConsensusCommitAdminIntegrationTestWithServer.java
#	server/src/integration-test/java/com/scalar/db/server/DistributedStorageAdminIntegrationTestWithServer.java
#	server/src/integration-test/java/com/scalar/db/server/ServerAdminTestUtils.java

+++ core/src/main/java/com/scalar/db/transaction/consensuscommit/ConsensusCommitAdmin.java
@@ -13,6 +13,7 @@ import com.scalar.db.config.DatabaseConfig;
 import com.scalar.db.exception.storage.ExecutionException;
 import com.scalar.db.io.DataType;
 import com.scalar.db.service.StorageFactory;
+import com.scalar.db.util.ScalarDbUtils;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
@@ -55,18 +56,30 @@ public class ConsensusCommitAdmin implements DistributedTransactionAdmin {
 
   @Override
   public void createCoordinatorTables(Map<String, String> options) throws ExecutionException {
+    if (coordinatorTablesExist()) {
+      throw new IllegalArgumentException("coordinator tables already exist");
+    }
+
     admin.createNamespace(coordinatorNamespace, options);
     admin.createTable(coordinatorNamespace, Coordinator.TABLE, Coordinator.TABLE_METADATA, options);
   }
 
   @Override
   public void dropCoordinatorTables() throws ExecutionException {
+    if (!coordinatorTablesExist()) {
+      throw new IllegalArgumentException("coordinator tables do not exist");
+    }
+
     admin.dropTable(coordinatorNamespace, Coordinator.TABLE);
     admin.dropNamespace(coordinatorNamespace);
   }
 
   @Override
   public void truncateCoordinatorTables() throws ExecutionException {
+    if (!coordinatorTablesExist()) {
+      throw new IllegalArgumentException("coordinator tables do not exist");
+    }
+
     admin.truncateTable(coordinatorNamespace, Coordinator.TABLE);
   }
 
@@ -166,6 +179,10 @@ public class ConsensusCommitAdmin implements DistributedTransactionAdmin {
       String namespace, String table, String columnName, DataType columnType)
       throws ExecutionException {
     TableMetadata tableMetadata = getTableMetadata(namespace, table);
+    if (tableMetadata == null) {
+      throw new IllegalArgumentException(
+          "Table does not exist: " + ScalarDbUtils.getFullTableName(namespace, table));
+    }
     String beforeColumnName = getBeforeImageColumnName(columnName, tableMetadata);
 
     admin.addNewColumnToTable(namespace, table, columnName, columnType);
